requirements.txt
Pillow
Numpy
opencv-python